requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
arabic-reshaper==3.0.0
    # via xhtml2pdf
asgiref==3.8.1
    # via
    #   django
    #   django-htmx
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   pyhanko
    #   pyhanko-certvalidator
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
chardet==5.2.0
    # via reportlab
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pyhanko
crispy-bootstrap5==2024.2
    # via -r requirements.in
cryptography==42.0.7
    # via
    #   pyhanko
    #   pyhanko-certvalidator
cssselect2==0.7.0
    # via svglib
dj-database-url==2.1.0
    # via environs
dj-email-url==1.0.6
    # via environs
django==5.0.6
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   dj-database-url
    #   django-allauth
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-extensions
    #   django-htmx
    #   django-taggit
    #   django-tinymce
    #   django-unfold
    #   djangorestframework
django-allauth==0.63.1
    # via -r requirements.in
django-cache-url==3.4.5
    # via environs
django-cleanup==8.1.0
    # via -r requirements.in
django-crispy-forms==2.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-debug-toolbar==4.3.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-htmx==1.17.3
    # via -r requirements.in
django-taggit==5.0.1
    # via -r requirements.in
django-tinymce==4.0.0
    # via -r requirements.in
django-unfold==0.29.1
    # via -r requirements.in
django-widget-tweaks==1.5.0
    # via -r requirements.in
djangorestframework==3.15.1
    # via -r requirements.in
environs[django]==11.0.0
    # via
    #   -r requirements.in
    #   environs
freetype-py==2.3.0
    # via
    #   reportlab
    #   rlpycairo
html5lib==1.1
    # via xhtml2pdf
idna==3.7
    # via requests
lxml==5.2.2
    # via svglib
marshmallow==3.21.2
    # via environs
oscrypto==1.3.0
    # via pyhanko-certvalidator
packaging==24.0
    # via marshmallow
pillow==10.3.0
    # via
    #   -r requirements.in
    #   reportlab
    #   xhtml2pdf
psycopg2==2.9.9
    # via -r requirements.in
pycairo==1.26.0
    # via rlpycairo
pycparser==2.22
    # via cffi
pyhanko==0.25.0
    # via xhtml2pdf
pyhanko-certvalidator==0.26.3
    # via
    #   pyhanko
    #   xhtml2pdf
pypdf==4.2.0
    # via xhtml2pdf
pypng==0.20220715.0
    # via qrcode
python-bidi==0.4.2
    # via xhtml2pdf
python-dotenv==1.0.1
    # via environs
pyyaml==6.0.1
    # via pyhanko
qrcode==7.4.2
    # via pyhanko
reportlab[pycairo]==4.0.9
    # via
    #   -r requirements.in
    #   reportlab
    #   svglib
    #   xhtml2pdf
requests==2.32.2
    # via
    #   pyhanko
    #   pyhanko-certvalidator
rlpycairo==0.3.0
    # via reportlab
six==1.16.0
    # via
    #   html5lib
    #   python-bidi
sqlparse==0.5.0
    # via
    #   django
    #   django-debug-toolbar
svglib==1.5.1
    # via xhtml2pdf
tinycss2==1.3.0
    # via
    #   cssselect2
    #   svglib
typing-extensions==4.11.0
    # via
    #   dj-database-url
    #   qrcode
tzlocal==5.2
    # via pyhanko
uritools==4.0.2
    # via pyhanko-certvalidator
urllib3==2.2.1
    # via requests
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
whitenoise==6.6.0
    # via -r requirements.in
xhtml2pdf==0.2.12
    # via -r requirements.in
